+++ plugins/voxelman/block/plugin.d
@@ -5,16 +5,22 @@ Authors: Andrey Penechko.
 */
 module voxelman.block.plugin;
 
-import std.array : Appender;
+import std.experimental.logger;
+import std.array : Appender, array;
+import cbor;
 import pluginlib;
 import voxelman.core.config : BlockId;
 import voxelman.storage.coordinates;
 import voxelman.block.utils;
+import voxelman.utils.mapping;
+
+import voxelman.net.plugin;
+import voxelman.world.plugin;
 
 shared static this()
 {
-	pluginRegistry.regClientPlugin(new BlockPlugin);
-	pluginRegistry.regServerPlugin(new BlockPlugin);
+	pluginRegistry.regClientPlugin(new BlockPluginClient);
+	pluginRegistry.regServerPlugin(new BlockPluginServer);
 }
 
 
@@ -29,13 +35,13 @@ struct BlockInfo
 	ubyte[3] color;
 	bool isVisible;
 	bool isTransparent;
+	size_t id;
 }
 
 final class BlockManager : IResourceManager
 {
 private:
-	BlockId[string] blockMap;
-	immutable(BlockInfo)[] blockInfos;
+	Mapping!BlockInfo blockMapping;
 
 public:
 	override string id() @property { return "voxelman.block.blockmanager"; }
@@ -48,23 +54,22 @@ public:
 	BlockId regBlock(string name, ubyte[3] color, bool isVisible,
 		bool isTransparent, Meshhandler meshHandler)
 	{
-		BlockId newId = cast(BlockId)blockInfos.length;
-		blockMap[name] = newId;
-		blockInfos ~= BlockInfo(name, meshHandler, color, isVisible, isTransparent);
-		return newId;
+		auto id = blockMapping.put(BlockInfo(name, meshHandler, color, isVisible, isTransparent));
+		assert(id <= BlockId.max);
+		return cast(BlockId)id;
 	}
 }
 
-final class BlockPlugin : IPlugin
+mixin template BlockPluginCommonImpl()
 {
 	private BlockManager bm;
 	// IPlugin stuff
 	mixin IdAndSemverFrom!(voxelman.block.plugininfo);
+	immutable string blockMappingKey = "voxelman.block.block_mapping";
 
 	override void registerResourceManagers(void delegate(IResourceManager) registerHandler)
 	{
-		bm = new BlockManager;
-		registerHandler(bm);
+		registerHandler(bm = new BlockManager);
 	}
 
 	override void registerResources(IResourceManagerRegistry resmanRegistry)
@@ -73,13 +78,67 @@ final class BlockPlugin : IPlugin
 		bm.regBlock("dirt", [120, 72, 0], true, false, &makeColoredBlockMesh);
 		bm.regBlock("stone", [128, 128, 128], true, false, &makeColoredBlockMesh);
 		bm.regBlock("sand", [225, 169, 95], true, false, &makeColoredBlockMesh);
+		registerResourcesImpl(resmanRegistry);
 	}
 
 	immutable(BlockInfo)[] getBlocks()
 	{
-		return bm.blockInfos;
+		return cast(typeof(return))bm.blockMapping.infoArray;
 	}
 }
 
+final class BlockPluginClient : IPlugin
+{
+	mixin BlockPluginCommonImpl;
 
+	override void init(IPluginManager pluginman)
+	{
+		auto connection = pluginman.getPlugin!NetClientPlugin;
+		connection.regIdMapHandler(blockMappingKey, &handleBlockMap);
+	}
 
+	void handleBlockMap(string[] blocks)
+	{
+		bm.blockMapping.setMapping(blocks);
+		infof("received block map");
+	}
+
+	void registerResourcesImpl(IResourceManagerRegistry resmanRegistry){}
+}
+
+final class BlockPluginServer : IPlugin
+{
+	mixin BlockPluginCommonImpl;
+
+	void registerResourcesImpl(IResourceManagerRegistry resmanRegistry)
+	{
+		auto ioman = resmanRegistry.getResourceManager!IoManager;
+		ioman.registerWorldLoadHandler(&handleWorldLoad);
+	}
+
+	override void init(IPluginManager pluginman)
+	{
+		auto connection = pluginman.getPlugin!NetServerPlugin;
+		connection.regIdMap(blockMappingKey, bm.blockMapping.nameRange.array);
+	}
+
+	void handleWorldLoad(WorldDb wdb) // Main thread
+	{
+		ubyte[] data = wdb.loadPerWorldData(blockMappingKey);
+		scope(exit) wdb.perWorldSelectStmt.reset();
+
+		if (data !is null)
+		{
+			string[] blocks = decodeCborSingleDup!(string[])(data);
+			bm.blockMapping.setMapping(blocks);
+		}
+
+		auto sink = wdb.tempBuffer;
+		size_t size = 0;
+		auto blockInfos = bm.blockMapping.infoArray;
+		size = encodeCborArrayHead(sink[], blockInfos.length);
+		foreach(info; blockInfos)
+			size += encodeCbor(sink[size..$], info.name);
+		wdb.savePerWorldData(blockMappingKey, sink[0..size]);
+	}
+}